feat(quantity): 商品数量全域整数化（V99）——23 列 NUMERIC→INTEGER，输入门禁收紧为正整数
依据：生产库实测 23 个数量列（含两个数量乘数）零小数存量；件数域本就应为整数
（SKU 主数据线新列已全部选 INTEGER）。方向指令来自用户 2026-08-31。

- V99 迁移：23 列 ALTER TYPE INTEGER（fail-loud 小数预检、触发器/生成列/analytics
  四视图先删后建，定义取自生产活库）；docs/schema.sql 快照与迁移历史契约同步 91 行
- 实体/DTO：8 个 JPA 实体数量字段 BigDecimal→Integer；OrderLine 族 DTO 数量改
  JSON 整数；乘数应用改 Math.multiplyExact，绝不静默取整
- 输入面：POSITIVE_DECIMAL_QUANTITY 退役，数量输入统一正整数 pattern；文件解析
  小数数量落 QUANTITY_SCALE 待复核；礼包行数量非整数在三路共用接缝抛
  BUNDLE_QUANTITY_NOT_INTEGER（SourceBundleResolver.requireIntegerQuantityForBundle）；
  回传实际数量校验改「正整数且不超指令数」
- raw-JDBC 边界：queryForMap 对 INTEGER 列返回 Integer，清理全部 (BigDecimal) 强转
- openapi：数量 schema 改整数文本/整数类型，价格与净含量保留十进制（新 DecimalAmount/
  PositiveDecimalAmount）；前端响应类型跟随（OrderLine 族 number）
- 测试：小数夹具按新语义改写或转拒绝断言；迁移测试快照数值语义归一化

+++ backend/src/main/java/cn/zimu/fulfillment/file/SourceOrderCandidate.java
@@ -2,7 +2,6 @@ package cn.zimu.fulfillment.file;
 
 import cn.zimu.fulfillment.common.audit.AuditActorType;
 import cn.zimu.fulfillment.order.dto.CanonicalOrderInput;
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -62,7 +61,7 @@ record SourceOrderCandidate(
             String sourceSkuRef,
             Long skuId,
             String skuCode,
-            BigDecimal quantityMultiplier) {
+            Integer quantityMultiplier) {
         SourceMappingSnapshot {
             if (itemIndex < 0
                     || (componentIndex != null && componentIndex < 0)
@@ -80,7 +79,7 @@ record SourceOrderCandidate(
             int partitionCount,
             String sourceBundleRef,
             Long bundleId,
-            BigDecimal quantityMultiplier,
+            Integer quantityMultiplier,
             List<BundleComponentSnapshot> components) {
         SourceBundleMappingSnapshot {
             if (rawImportRowId <= 0
@@ -96,7 +95,7 @@ record SourceOrderCandidate(
     }
 
     /** 静态礼包 BOM 的内部 SKU 身份与每礼包数量。 */
-    record BundleComponentSnapshot(Long skuId, String skuCode, BigDecimal quantityPerBundle) {
+    record BundleComponentSnapshot(Long skuId, String skuCode, Integer quantityPerBundle) {
         BundleComponentSnapshot {
             if (skuId == null || skuCode == null || skuCode.isBlank() || quantityPerBundle == null) {
                 throw new IllegalArgumentException("礼包组件快照不完整");
